Start the parallelization at the outmost loop
This makes a real difference with ICC's OpenMP. Without this patch, the
plugin's total time is about 15% worse compared to GCC 4.8, and time
reported as spent in all CPUs is worse by 19%. According to the output
of `perf top`, it seems that about 5% of time was being spent dealing
with barriers (__kmp_hyper_barrier_release etc).

GCC's times are roughly the same as before.