There's no "username" in new Facebook APIs
The "username" got removed somewhere in the 2.x Graph API version
history. Given that we don't actually need it for anything but showing
the current account to the user, let's just replace it with the
human-readable name of the current account.

(cherry picked from commit 075bfdcad84beba534edef04bf989470be542a7a)