Enhance module unscrollability.
NOTE: this makes side-panels very greedy in hijacking && ignoring
scrolling - now you will only be able to scroll side panels with scrollbars.

I think this is better than we have now and what we had in 1.6.x,
though if there will be significant amount of complaints we might
revert and push a less greedy version.