Demosaic IOP: new method: passthrough (monochrome)
For CFA cameras with removed CFA filter.

Basic idea is similar as how we treat Leica M Monochrom:
There is 1 channel, and we load it into each of 3 channels.

Since we can distinguish that camera at raw loading time,
in that case we do it at raw loading time, thus, for our pipe
images from that camera are just pre-demosaiced raws.

But in case of CFA cameras with removed CFA filter, no metadata
is changed, so the only way to handle this is to add special
demosaic mode, that will propagate values from the single channel
into all 3 channel, thus, all 3 channels will contain the same data.