Demosaic IOP: process(): if method = monochrome, do not switch to other methods
If in DR mode we request very small-quality image, for speed reasons
we fallback to either PPG or MARKESTEIJN method.

But. If method is monochrome, there would be significant difference
between image processed with it, and any other method, also it is
extremely fast, so we should not switch to other methods in this case.