Fix num values when importing pictures.
Another instance of this issue! The num in history must start to 0 and
increment one by one without gap. The trick of using a memory table and
use the rowid does not always work. In this case the statement is complex
enough to "confuse" sqlite and the rowid becomes "random". The added
code then recompute the num properly to order them as expected in the
code inserting the items into the history table.

Hopefully this is the last instance of this!