clip&rotate: fix bug in flip function, fix for #10511
Due to rounding issues the old flipping code would lead to an undefined (1px black) line at
one or two of the image borders - the image is shifted by one pixel relative to its buffer.
With this fix images are now correctly positioned in the buffer, however, this leads to a
-1px offset versus the old code. I would call it a bug fix without the need to deal with
legacy history stacks.
and/or bottom border