Do not create multi-instance history items for single-instance modules
The cleanup code introduced due to bug #10192 in commit
0afe981dc5c1d24a1e9867386eb1140382fff363 (thanks to LebedevRI for
tracking it down). If the end of the history stack contained data for
the same IOP module as the one we were pasting, the pasted instance got
turned into a multi-instance one by accident.

This patch preserves creation of extra history rows for these pasted
modules. I suppose that this is how it is supposed to work; users are
probably expected to compress the history stack afterwards if they don't
want to have N more items after each paste of N iops which were already
present in the stack. That's how I recall DT ever working, anyway.

Thanks to LebedevRI and houz for hand-holding me through DT's internals.

fixes #10539