Grab focus of main area when the mouse enter it.
After clicking on the left panel or rigth panel, the next key-events
are redirected to those panels. For example after selecting a metadata
entry, moving over the lightable and pusing 'z' (full preview) won't
work. Actually the 'z' is enterred into the metadata entry. The same
applies to all actions on the side panels.

Fixes #10508