get_matrix_from_profile(): if there is a LUT, give up. Fixes #10552
If profile contains only a LUT, we would fall-back to LCMS codepath.
If profile contains a matrix + LUT, we would just use matrix in our
fast matrix codepath.

But, if profile contains a LUT, it might contain a swapped matrix, so
some apps (like us right now) would use matrix, and get an image with
swapped colors, and some would use LUT and get image with right colors.

The matrix in a XYZLUT profile is intended for older outdated CMS
implementations and should not be used.

Falling back to matrix used to kinda-sorta make sense when our fastpath
was an order of magnitude faster than our LCMS2 path,
which is no longer true. (pr #495 + pr #509)