gpx: parse only elements within trk
I came to use a few KML<->GPX online converters, and also a popular
open source gps GPX logger on Android. These were generating GPX files
that darktable would throw a warning at, and the message is not very
clear. Try to avoid that.

The basic document structure was:
<gpx>
<wpt lat="xxx" lon="yyy">
 <ele>average of the waypoints</ele>
 <name>Description</name>
</wpt>
<trk>
 <trkpt></trkpt>...
</trk>
</gpx>

This structure is perfectly legit when looking at the XML schema file.

The problem is that we parse all "wptType" sub elements wether or not we're
in a trk>trkpt or not, and DT chokes if no trkpt is actually being parsed.

So we must better track what we're actually parsing and currently limit the
wptType subelements parsing within trk>trkpt