Revert "rawspeed: add bitdepth modes to SRW"
This reverts commit 70a60f14581ff66448e0584a70def8a439e38883.

It seems the sample files provided for the NX500 actually use some decoding features that were not implemented yet. At least some of them seem to be related to this bitdepth stuff so it may not even be required to do 12bit and 14bit cameras.xml modes for the camera. This code also broke all other SRW cameras to this needs some deeper work.