Preload images in the full-size preview and simplify SQL a bit
The previous version of these SQL statements were a result of a few
years of incremental changes; under certain circumstances it was doing a
subquery and joining the result back with the original table. This
change simplifies the logic a bit -- the patch is best viewed when
ignoring indentation.

The real motivation was however speculative preload of a couple of next
images. Loading them in background cuts the latency when the user
navigates images that have expensive IOPs in theit history stacks. See
http://thread.gmane.org/gmane.comp.graphics.darktable.user/7832 for
details.

The number of images to preload is configurable (not through the GUI,
though). Right now, the code only preloads one image as a conservative
default; this is a result of a GitHub/IRC discussion.

Because the job queue is not a queue but a stack, we have to reverse the
list of numbers returned from SQL in the C code (or use a subquery,
but none of these is terribly elegant IMHO). Thanks to houz, LebedevRI
and mgehre for their feedback.