mipmap_cache: DT_MIPMAP_BEST_EFFORT: only load MIP_0 if it exists on disc
The current check is wrong, and only loads/creates the MIP_0 if it does
not exist on the disc. This is opposite to what the comment above says
and also increases loading times.
If no MIP for that image exists, DT_MIPMAP_BEST_EFFORT should only
create the requested MIP and no other levels before.