Hopefully fix some more crashes
The problem is that some code (dt-cli, init of printing) runs a pipe and
using the display profile before such a profile could be fetched from
the system. So we are initializing the system display profile with sRGB
and hope that no one really relies on it being correct at that stage.