dt_{mipmap,}_cache_get(): check that we are not returning 'w' lock when asked for 'r'.
When entry is not found by dt_cache_get(), new entry is allocated, and
in most cases (when asked for 'w' lock OR when global cache->allocate
callback is set) 'w' lock is returned.

We do not want to keep 'w' lock unless we have asked for it,
so additional logic for lock demoting is needed.