Generate-cache: cleanup/rewrite. Influenced by PR #970
Re-uses (deduplicates) the logic inside the mipmap cache for scaling
and writing thumbnails.

Also, write each image thumbs once they are generated, and not when
they are gc'd or g-c finishes.
This way, if user kills g-c in the middle of the process, more thumbs
will be on the disc.

TODO: multi-threaded g-c?