Add keyboard shortcuts to control the brush hardness and opacity.
When using a tablet and a stylus you don't always have a way to scroll
up and down. So we are forced to get back to the mouse to change brush
hardness and opacity. This patch adds four key accels : { and } are
standards shortcuts to change the brush hardness. < and > are also
standards in Gimp to change the brush opacity.

Of course, now it's also possible to map the tablets buttons to these
keys.