Revert a6e5655f and edb299d1
This patch to fix a warning reported by CMake when Ninja is used as the
build tool incorrectly reverted 91ed7438. Using an absolute path for
the DTD in the XML source is not possible because of the possibility of
non-ASCII characters in the path.

A new patch will be prepared later.