Switched back to normal Lensfun search mode for lens lookups.
The option LF_SEARCH_SORT_AND_UNIQUIFY was needed when darktable checked for
uniqueness of the search result (see bug #9662).  However, since commit
362ba6a, these checks have been removed, and now, LF_SEARCH_SORT_AND_UNIQUIFY
is harmful in case of lookups.  The reason is that in contrast to what commit
message 362ba6a says, the exact lens model name may not be the very first
returned element.