tools: camera support compare without digests This now makes it quite easy to generate the camera support changelot to use in new releases. A few examples:
To get the current diff from 2.0rc3:
$ ./tools/check_camera_support --compare release-2.0rc3 master

And to get the current diff from 1.6.9:
$ ./tools/check_camera_support --compare 24e02e8a master

The second example is not just "1.6.9 master" because the script doesn't work with the file formats used in 1.6.x so I had to find a commit that the same camera support as 1.6.9 did but in master. That's only an issue for this release anyway.