tools: first import of regression testing scripts
This is the code that drives http://darktable.pedrocr.pt/
It is a way of testing darktable against itself to make
sure we are not introducing bugs that break history stacks.
To do this it compiles specific versions and uses dt-cli
to check that the output of one is the same as the other.
XMPs will be added to the repo and for raw files the
rawsamples.ch collection will be used.