tools: add a simple runner script for everything
This marks a first easily usable version of the regression
testing stuff. Running it is pretty simple, just do:

$ ./compare darktable-1.6.x master

This will fetch all the samples from rawsamples.ch (long
download and 8GB+ of space), fetch and compile the two
versions of darktable, and then proceed to run the comparison
between the two versions and generate the html files to
describe it. On subsequent runs darktable will only be
recompiled if there are new commits and only new samples
are downloaded.