tools: group cameras by make and model
Using cameras.xml and the exif info of each file we can
group the cameras being tested under their clean make and
model names. This makes the output cleaner and also allows
things like:

$ ./compare darktable-1.6.x master canon s100

which will run the comparison only for the Canon S100. The
match is case insensitive to make it easier to specify on
the command line.