dt_dev_pixelpipe_get_dimensions(): properly handle skipped modules.
Fixes #10792.

Also, this *might* fix those strange tiling memcpy() crashes,
but i'm failing to find an issue in redmine, and i was never
able to reproduce that crash...

The problem here was that if for all the pieces that if statement
computes false, we never set roi_out, so after the loop it
still contains same garbage (it was not even zero-initialized).
And then we use that garbage to init *width and *height...