demosaic: support VNG4 demosaic of CYGM cameras
Bayer sensors with CYGM filters were not supported because
the demosaic output needs to be 4 channel so that colorin
can then use a 4x3 matrix to get to RGB. As it turns out
the VNG4 interpolation we got from dcraw already supports
these just fine we had just crippled it with some pre and
post processing that assumes g1==g2. After this commit the
Canon G1 and Nikon E5700 seem to work decently. The dcraw
output seems less noisy though.