temperature: 4color temperature correction
CYGM images need four independent WB scale factors.
Practically all the code in demosaic follows the same
pattern as dcraw but does the slight simplification of
doing 3x3 instead of 4x3. This adds back that 4x3 case
which in general is just the same code extended but
now the pseudo matrix inverse code is needed as inverting
a non-square matrix is not mathematically possible. In the
normal 3x3 case this should be exactly equivalent.