temperature: always do CYGM->RGB after WB
WB is supposed to happen on original raw values but in the
preview case the MIPF is already demosaiced. In the CYGM case
that can be worked around by instead of doing CYGM->RGB on
MIPF creation, create the MIPF as CYGM and convert to RGB
in demosaic itself after applying the WB. This assumes it's
safe to keep the extra channel in the MIPF which is fine
as long as no iop before temperature that is enabled in the
preview pipe allows masks.