fix off-by-two error when seeding v5sum
This caused slightly wrong pixels in the first few columns of each
tile. Thank you to Ulrich Pegelow for noticing that there was a regular
grid of deltas between the results of his OpenCL code and the CPU-only
code at TS intervals. The code upegelow wrote is correct, and formed a
reference to find the bug in the (original) CPU-only code.