simplify initializing 5x5 homogeneity maps step
Rather than pre-calculate v5sum (rolling sum of last five columns) based
upon the five columns before the first column processed, start
processing at a column at which we know v5sum will be 0 due to the 0
padding at the edges of the homogeneity maps. This simplifies the code,
may be faster, and leaves less room for off-by-two bugs such as the one
fixed in the previous commit.