colorspaces: OpenMP for CYGM<->RGB conversions
At least dt_colorspaces_cygm_to_rgb gets run for the full
image so can use OpenMP to speed it up. The RGB->CYGM
transform is only ever called for num==1 so this should have
no effect but in case we end up using it for something
heavier in the future it's already sped up.