temperature: use the embedded d65 matrix for temp
To do the Kelvin->Multipliers and Multipliers->Kelvin
conversion we use XYZ_to_CAM and CAM_to_XYZ matrices
from adobe_coeffs. For DNGs with embedded matrices
these don't exist but we have the matrix in dt_image_t.
Use that whenever it is set as we also default to that
matrix in colorin if it exists. (fixes #10899)