histogram: Corrected mouse scrolling/dragging on the histogram (fixes #10073 and #10705)
Mouse scrolling and dragging on the right part of the histogram (exposure) was
implemented in terms of "white", not in terms of "exposure". That caused
incorrect behavior (#10705): the slightest drag in any direction changed
exposure by several EV, scrolling changed exposure in non-linear way. On the
other hand dragging and scrolling in the "auto" exposure mode was done
in terms of target exposure value, passing "white" and "exposure" via
the same hooks caused #10073.