imageop: make the filter cropping recalculatable
Instead of calculating the filter cropping only once at load
make it a function we can call at any time but still reusing
the rawspeed code. This way when in the future we start saving
the crops to the history stack we already have the code in
place to be able to recalculate whatever crop we need.