invert: make it work properly for CYGM images
Invert uses an RGB color as params but then works on
pre-demosaic data. When we have a CYGM image convert
that RGB point to CYGM and then apply the transformation
as normal. This actually fixes an out of bounds access.
FC() returns 0-3 for these images which would then be
used to index a 3-wide array. The preview pipe already
worked well because that's already RGB since init_f and
that's unchanged here.