rawspeed: accept SubIFDs that use type 13
Some formats have decided to use TIFF type 13 to
mean "offset within the file" to make it easier to
rewrite TIFFs. Recently some DNG files created by
the x3f_extract tool showed up and were broken. As
it turns out it was just a SubIFD being ignored
because it was using the type 13 data type for the
offset and although getInt() had already been
changed to accept it getIntArray() hadn't. Fix that
by accepting the type 13 as well.