demosaic: implement monochrome for MIPF
Simplify monochrome demosaic by handling the MIPF
case in demosaic.c as well. When we get to demosaic
we have an RGB image so we just set R, G and B all
to average(R,G,G,B). Since the image is small and
it's been half-size demosaiced anyway this doesn't
produce any visible artifacts. This will also do
away with the need for the extra complexity in
init_f().