rawspeed: take into account FileMap extra space
FileMap includes by default 16 bytes of extra space so that
things like BitPumpMSB can read safely a bit beyond the size
of the file. Take that into account properly when doing bounds
checking. Fixes some broken D800 files that had the image data
go all the way to the end of the file.