ashift: additional dummy parameter "toggle" for module activation
the module needs to be active else there are no data for structure
extraction. therefore on first click of the "structure" or "fit" buttons
a warning is displayed. "toggle" makes sure that the module gets
activated now (even if no other parameters have changed). on the next
click of one of the buttons data should be available.