rawspeed: Correctly handle too large TIFF entries
TIFF entry count of items is a 32bit integer but the
corresponding byte size can overflow those 32bits if the
count is very large and the byte size is 4 or 8. Reject
the tiff entry if the byte size exceeds UINT32_MAX as
we don't handle files larger than that anyway.