demosaic: take advantage of new FCxtrans() code
Now that FCxtrans can take a roi and already has an offset of 6, can
simplify Markesteijn demosaic code accordingly.

Note that offset for mirroring a border pixel is reduced from 24 to 12,
as row/col will always be >= -17, and there is already a +6 offset in
FCxtrans(). It had been set unnecessarily high.