demosaic & imageop_math: increase FCxtrans offset
Increase from 6 to 18 (must stay a multiple of 6) so that don't have to
use additional harcoded offsets in CPU-based Markesteijn demosaic. So
long as there's an addition already in FCxtrans(), this shouldn't alter
performance, so might as well keep the offset in one place. An offset of
18 is enough to handle the most negative values from Markesteijn
demosaic, and more than enough for VNG and the non-demosaic iops.