ashift: add "sweeping" mode for easier selecting/deselecting of lines
keep the button pressed and move the mouse to select and deselect lines
as you go. the cursor shape reflects the mode you are in. if you are zoomed
into the image sweeping mode is only entered if you are close to a line
when mouse button is clicked, else you are in normal image panning mode.