tools: handle Pentax naming in gen-profile
Newer Pentax cameras have "Ricoh" as Maker and "Pentax <model>"
as the model name. Set the make/model to "Pentax"/"<model>" in
those cases to match what we actually use.

The full fix for this would be to just use the exif make/model
and do a lookup in cameras.xml. There's Ruby code to do that in
tools/check_camera_support but it would require either depending
on ruby for gen-profile or doing something equivalent in shell.

Fixes #10925