IOPs: add #include "iop/iop_api.h"
Yes, we need to include it via cmake AND manually.
Via CMake - in case it is not included manually.
Manually - cmake && make does not detect when the header
included via '-include' is changed and does not cause
recompilation, only #include "..." does.

Also, not all (not mine) IDE understand -include,
so they don't parse that header => not cool for reasons.

(to check whether all iops include the header one can use
$ grep -lr DT_MODULE * | xargs grep -L "iop/iop_api.h")