x-trans: generalize handling negative CFA offsets
Up until now FCxtrans() guarded against a negative offset into the CFA
by adding a constant to incoming row & col, with the constant based on
knowledge of the most negative row & col used in demosaic
algorithms. But this could cause a bug if a future use of FCxtrans()
involves unexpectedly negative inputs.

This commit defines a modulo function that always returns a positive
value (so long it is given a positive divisor). Both FCxtrans() and the
HEXMAP() macro use this function. The downside is that this change
results in approx. 4% slower processing of an X-Trans image when using a
minimal set of iops.