tools: improve make/model detection
Instead of returning the rs-identify make/model directly
still pass it through the exif->clean_name map to make
sure we get the best possible name for the camera. We could
just use the canonical_make/canonical_alias from identify
but that may not be using the same cameras.xml file.