rawspeed: avoid NULL-derefs in Sony A100/F828
getEntryRecursive() has a tricky API compared to getEntry()
because it will happily return NULL if it doesn't find
anything, whereas getEntry() will throw an exception. This
makes the getEntryRecursive()->getInt() pattern unworkable.
For now fix these two cases by first searching for the IFD
and then using getEntry() but in the future we should probably
consider a more uniform API.