Show accel in button tooltips & some accel cleanup
We already mentioned some accels in the tooltips but hardcoded them to
the initial default values (and in the case of invert-selection it was
the wrong one) so it didn't reflect the actual user settings. Now the
right keys should be shown.
I had to move some accels to other libs and hope that it didn't break
anything.