Revert "liquify: gui_post_expose should distort points through all iop."
This reverts commit 038ed96b20232bfe35035cc6d74d156431b48ab9.

The proper fix would be to distort the controls through all modules
except liquify. We do not want the control to be distorted, the control
must stays at the position it has been placed by the user. Note that
the gui_post_expose and the get_point_scale routines must distort the
points through the same modules.

For #10939