dt_dev_distort_[back]transform_plus add exception for module.
In these routines we had a way to select the range of module priority
to use to apply the distortion. A new parameter except is introduced
to be able to add an exception for a module to skip during the
computation. If the parameter is NULL there is no exception and
the routines are behaving as before.

This is needed and used in liquify as the distortion for the control
elements must be computed for the whole pipe (before liquify we have
lens and ashift to take into account, and after roration and crop)
except liquify itself which must not modify the control point.

Proper fix for #10939.