demosaic: x-trans clamping to process_maximum
Assert that incoming Markesteijn data is [0,process_maximum]. Then
remove vestigial clamping to [0,1] of in-between steps and output (this
is a throwback to dcraw's integer pipeline), but still clamp the output
to [0,process_maximum]. Similarly clamp VNG demosaic output (which will
apply to all sensor types).

As a fix-up, change some constants in Markesteijn equations from int to
float.