tonecurve: add second automatic colour scaling option.
this has the same effect on saturation as rgb curves would.  it is done by
applying the same curve that is draw for L to X,Y and Z channels instead of
doing it in Lab. note that this is done in a way that preserves the
L_out = C(L_in) behaviour, just the colour channels are affected.