rawspeed: TiffEntry::getInt() for arrays
getIntArray() and friends are a nasty interface because they can
very easily lead to out of bounds access if the caller isn't careful.
Instead generalize getInt() and friends to take an array index and
thus be a full replacement for getIntArray() with additional bounds
checking. Since these are used for metadata and other low volume
usage performance impact should be negligible.