ashift: skip interpolation step in case of all-neutral parameter settings
if all parameters are zero we just copy input->output. as this comes with
almost no performance penalty it is now inexpensive for users to make this
module enabled by default. advantage: data for structural analysis is present from
the beginning, making perspective correction more or less a one-click option.