Exposure iop: compute_correction(): use sep blacks. Refs #10995, Refs #10996.
For some cameras rawspeed provides some really bogus value
for black level, e.g for  PENTAX K-5 it is reported as 512,
while sep blacks are calculated at ~67.

For deflickering it is really unfortunate.

Since this is the only place where we use black level,
and everywhere else we use sep blacks, let's just
average those. If sep blacks are wrong - we will notice.

Note: this technically does change the output,
but only in cases where rawspeed is lying.