highlights: x-trans Lch without edge noise
Reconstruct via 3x3 samples centered on the current pixel. But don't
reconstruct a pixel if it adjoins a 3x3 sample which is unclipped. For
example, if the current pixel is in the bottom-left corner of an
unclipped 3x3 area but there are clipped pixels below and to its left,
don't reconstruct it.

This fixes zippering in sharp transtions from clipped to unclipped areas
caused by the irregular pattern of the X-Trans sensor.

The results should be as good at edges as the prior X-Trans Lch
reconstruct algorithm, while taking advantage of the vastly better
results from the new Bayer Lch reconstruct algorithm.