Specify cache dimensions and image ID range for cache regen
In my workflow, I tend to rely on the previews quite heavily, both in
the zoomed-out lighttable viewing as well as in the zoomed-in
single-image lighttable mode where just one image is visible. I also
prefer computers to do boring tasks in advance, so I do not want to keep
waiting for these thumbnails to appear while I'm browsing them in the
GUI.

This isn't a perfect patch because a good one would essentially allow to
specify arbitrary combinations of the cache sizes (for example, I
usually need "5" and "2" on my system). However, this is easy enough to
change for now, and matches the feedback on #darktable I got (thanks!).

There's quite some mixing of int32_t vs. uint32_t going on when
referring to image IDs. The existing code (dt_image_t for example) uses
int32_t for the image ID anyway, and this way the atoi/max/min combo
works.