Fix non-working libsecret pwstorage backend
It seems this code was broken forever, resulting in it not even trying
to store passwords on disk but only in the session keyring. The result
was that people had to re-enter their passwords after every reboot.

(cherry picked from commit a423f36794b6479eb29f9f0168064feb5dacc84b)