Fix softproofing with some internal profiles
The problem is that lcms2 is too smart and not forcing clipping during
softproofing. That might be useful when simulating the effect the
profile alone has, but useless when simulating export to a ldr format.
The workaround is to dump the profile to memory and read it back so that
lcms2 quantizes the TRC.
See https://sourceforge.net/p/lcms/mailman/message/35178784/

(cherry picked from commit b6eb53bbba4310e517aaad5d6c3013d53522904e)