opencl: do not alias readonly with writeonly image buffers (colormapping)
this is a first of a series of required changes to some of our opencl
code. in several places we circumvent the ocl-1.x limitation to offer only
readonly and writeonly image objects by using aliases for the same buffer.
according to the ocl-1.2 standard this is no longer allowed and effectively
seems to lead to corrupted output in some opencl implementations (namely not in
NVIDIA and AMD drivers).