hotpixels: fix out of sync column for x-trans
Make initial column in loop variable be 2, to correspond to image
pointers which start at that column. This causes hotpixels to
compare X-Trans pixels to neighbors of the same color.

This fixes a bug reported by Marc Cousin. In his example image, the
hotpixels iop wrongly identified red pixels in a blurry red area as hot
green pixels, due to colors read from the wrong column of the X-Trans
CFA.

Note that this bug fix may change the needed threshold/strength
values for X-Trans images already processed via darktable, as the iop
now will compare pixels with a different set of neighbors.

While fixing this, have also (similar to Bayer code) removed checking
for hot pixels in a strip at the very right and bottom of the image,
This eliminates the need for code to check for out-of-bounds offsets. If
the pixel is at the image edge, there probably won't be 4 (or even 3)
neighboring pixels from which to know if it is hot, hence this wasn't a
useful code path.

Also, for clarity, added parens and comments to the offset table
generation.