Manual section 3.4.4.2: Fix spelling error and clarify slightly
"frequence band" should be "frequency band".

Also, the paragraph in question appears to be referring to the three
tabs, but it only refers to the final tab by its name ("edges"), while
the first two as lightness and chromaticity, which is not how the UI
refers to them. So, I added a reference to them as "luma" and
"chroma" to hopefully clarify this.