If there is no local-copy there is nothing to sync.
Check whether the image has a local copy or not. If not we exit as
there is really nothing to sync and so we avoid a spurious error
reported when an image has been removed/renamed outside of dt.

For #10725.

(cherry picked from commit 1f5cfff4f1a3eaed9efa10749a7f052cd1c25b3e)