config.h: add macros for manual ASan's memory (un)poisoning
In some places we keep buffers and reuse them to not
re-allocate each time. While this does save us from
malloc(3) overhead, this also shadows the actual
size that we think is usable from AddressSanitizer.

I.e. we first allocate buf of size 4, use it,
and then re-use it but with size 2.
And if there is some code that oveflows it, we
may not know about it.