Pixelpipe cache: manually (un)poison memory
We keep buffers and reuse them to not re-allocate
each time. While this does save us from malloc(3)
overhead, this also shadows the actual size that
e think is usable from AddressSanitizer.

I.e. we first allocate buf of size 4, use it,
and then re-use it but with size 2.
And if there is some code that overflows it, we
may not know about it.

With this hacky manual (un)poisoning, hopefully,
it will be easier to catch buffer overflows.