create_release.sh: manually do XZ compression.
I did compare every compression setting with the
tar-produced .tar.xz, and manually compressing with
options -9e results in the smallest archive,
with the 80% size of the one produced by tar directly.

This of course comes with speed and memory (de)compression
penalty, but since the archive is created just once,
it should be fine.