[tags] Allow to import/export tags
This is useful when using a closed vocabulary for tagging images. A way
to disable creating new tags is needed still, but this at least allows
to get tags into dt.

There are some free tag lists available on the internet, for example

- Some CC-BY-SA licensed lists:
  https://lightroom-keyword-list-project.blogspot.de/
- Not sure about the license:
  http://workshops.bigsunphotography.com/news/28-free-lightroom-keyword-library-download

Google might find more. Just make sure that the txt file uses the
Lightroom format as described here:
http://www.photo-keywords.com/keyword-list-format.php