Notify user when import/export is done
Tell the window manager to draw the user's attention to dt when an
import of a folder or an export is finished. Those two can take a long
time to be processed so the user might want to do something else outside
of dt in the meantime. Instead of listening to the CPU fan slowing down
as an indication he will now be told by his window manager/desktop
environment. On KDE for example the entry in the taskbar is highlighted.