CMake: just link to lib_darktable, not it's deps too.
Library dependencies are transitive by default with
this signature. When this target is linked into another
target then the libraries linked to this target will
appear on the link line for the other target too.

https://cmake.org/cmake/help/v3.0/command/target_link_libraries.html

One more reason why this is the right change:
$ cmake --graphviz=test.dot ..
$ dot -Tsvg test.dot -v > /tmp/cmake_ninja.svg

Before this, it was 'every single target depends on every single lib'
I.e. practicaly n^2.

Now, every single target (except a few stand-alone cli tools)
only depend on lib_darktable; and only lib_darktable depends
on libs. So it is kinda-linear.

This may somewhat affect the cmake/make timings...