tagging: bring back tag names to library.db
This introduces a table in the library database with id->name mappings.
On startup this data and tags / tagged_images is all synced so that at
runtime the new used_tags table isn't queried, just updated when tags
are changed.

As always, make a backup before trying this the first time.